% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshapeCP.R
\name{reshapeCP}
\alias{reshapeCP}
\title{A function to reshape form data from ChangePond (Best Plus, EPT)}
\usage{
reshapeCP(form, NitemsInPanel, NitemsInModule, outName = NULL)
}
\arguments{
\item{form}{Data frame of messy data downloaded from ChangePond}

\item{NitemsInPanel}{vector indicating the number of items in each panel, in the order that panels appear}

\item{NitemsInModule}{vector indicating the number of items in each module, in the order that modules appear in a panel}

\item{outName}{output file name to be saved as .csv in working directory}
}
\description{
This function reshapes form data downloaded from ChangePond
}
\examples{
###7 values for 7 panels. Each value represents the number of items in each panel.
ItemsInPanelsEPTS = c(3,3,4,5,5,5,10) 
#18 values for 18 modules. Each value represents the number of items in each module.
ItemsInModulesEPTS = c(3,3,2,2,1,2,2,1,2,2,1,2,2,2,2,2,2,2) 
#EPTform is the data frame that is downloaded from CP and read in as a data frame.
reshapeCP(EPTform, ItemsInPanelsEPTS, ItemsInModulesEPTS, outName = EPTform_clean)
}
\keyword{reshape}
